pyproject.toml
[project]
name = "djangorest-starter-kit"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Lugachev",email = "nikitalugachev149th@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
